refactor: integrate BridgeServer into main service + rewrite relay for Bun
- Mount BridgeServer routes at /browser/ on main Hono server (port 4096)
- Rewrite Extension relay from Node.js ws to Bun-native upgradeWebSocket
- Add process-level BridgeServer singleton (setBridgeServer/getBridgeServer)
- Update Chrome Extension default relay URL to ws://127.0.0.1:4096/browser/extension
- Remove standalone MCP server, HTTP client, and old extension-relay
- Add browser config schema with cdpPort/autoLaunch/headless options
