+++ opencode/packages/opencode/src/server/server.ts
@@ -44,6 +44,7 @@ import { PermissionRoutes } from "./routes/permission"
 import { GlobalRoutes } from "./routes/global"
 import { PreferencesRoutes } from "./routes/preferences"
 import { AgentTerminalRoutes } from "./routes/agent-terminal"
+import { BrowserRoutes } from "./routes/browser"
 import { MDNS } from "./mdns"
 
 // @ts-ignore This global is needed to prevent ai-sdk from logging warnings to stdout https://github.com/vercel/ai/blob/2dc67e0ef538307f21368db32d5a12345d98831b/packages/ai/src/logger/log-warnings.ts#L85
@@ -176,6 +177,7 @@ export namespace Server {
         .use(validator("query", z.object({ directory: z.string().optional() })))
         .route("/project", ProjectRoutes())
         .route("/pty", PtyRoutes())
+        .route("/browser", BrowserRoutes())
         .route("/config", ConfigRoutes())
         .route("/experimental", ExperimentalRoutes())
         .route("/session", SessionRoutes())
