+++ packages/browser-mcp-server/src/core/index.ts
@@ -1,4 +1,3 @@
 export * from './cdp'
 export * from './chrome'
-export * from './extension-relay'
 export * from './types'
